start work on new qrcodemodal to handle new unique mobile wallet connect qr code, update mobile direct connect qr code to be custom
