+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@provenanceio/walletconnect-js",
-  "version": "3.7.1",
+  "version": "3.7.2-develop.1",
   "private": false,
   "sideEffects": false,
   "main": "esm/index.js",
