+++ src/services/methods/connect/connect.ts
@@ -128,12 +128,16 @@ export const connect = ({
           connectionWalletAppId === 'figure_mobile_test';
         // New QR Codes will have three parts, WC Link, figure connect link, and dynamic link
         const wcLink = `${data}${requiredIndividualAddressParam}${requiredGroupAddressParam}${prohibitGroupsParam}${jwtExpirationParam}${connectionDurationParam}`;
-        const figureConnectLink = `https://figure.com/connect?link=${wcLink}`;
+        const figureConnectLink = `${encodeURIComponent(
+          'https://figure.com/wallet/connect?data='
+        )}${encodeURIComponent(encodeURIComponent(encodeURIComponent(wcLink)))}`;
         // TODO: Let dApp pass option here
-        const walletType:
-          | 'com.figure.mobile.wallet.dev'
-          | 'com.figure.mobile.wallet' = 'com.figure.mobile.wallet.dev';
-        const finalDynamicQRLink = `https://figurewallet.page.link/?link=${figureConnectLink}&apn=${walletType}`;
+        const mobileWalletDev = true;
+        const walletType = mobileWalletDev
+          ? 'com.figure.mobile.wallet.dev'
+          : 'com.figure.mobile.wallet';
+        const isi = mobileWalletDev ? 6444293331 : 6444263900;
+        const finalDynamicQRLink = `https://figurewallet.page.link?link=${figureConnectLink}&apn=${walletType}&ibi=${walletType}&isi=${isi}&efr=${1}`;
         const finalQRCodeLink = isMobileWalletOpened ? finalDynamicQRLink : wcLink;
         const qrCodeImage = await createQRImage(finalQRCodeLink);
         // Don't trigger a QRCodeModal popup if they say "noPopup" or pass a specific walletId
